Simplify paste integration
Simplify the paste integration.

The `--paste` option still works, but requires `plaster` and does not
support parsing the server section of the config file. Instead, server
configuration must be done with Gunicorn configuration files, command
line switches, and environment variables.

The use of Gunicorn as a paste server factory still works. It allows
specifying `host` and `port` or `listen` options, as well as `bind`, but
is otherwise more strict with options. Rather than ignoring unknown
options it will raise an error. There is no support for using a Gunicorn
config file when using a paste runner. Instead, all options must be
specified in the paste configuration file. No manipulation of the system
module path is performed.